Change uint to uint256

+++ contracts/lib/protobuf/InstrumentData.sol
@@ -21,7 +21,7 @@ library InstrumentParameters {
    * @return The decoded struct
    */
   function decode(bytes memory bs) internal pure returns (Data memory) {
-    (Data memory x,) = _decode(32, bs, bs.length);
+    (Data memory x, ) = _decode(32, bs, bs.length);
     return x;
   }
 
@@ -31,7 +31,7 @@ library InstrumentParameters {
    * @param bs The bytes array to be decoded
    */
   function decode(Data storage self, bytes memory bs) internal {
-    (Data memory x,) = _decode(32, bs, bs.length);
+    (Data memory x, ) = _decode(32, bs, bs.length);
     store(x, self);
   }
   // inner decoder
@@ -44,52 +44,55 @@ library InstrumentParameters {
    * @return The decoded struct
    * @return The number of bytes decoded
    */
-  function _decode(uint p, bytes memory bs, uint sz)
-      internal pure returns (Data memory, uint) {
+  function _decode(uint256 p, bytes memory bs, uint256 sz)
+    internal 
+    pure 
+    returns (Data memory, uint) 
+  {
     Data memory r;
     uint[6] memory counters;
-    uint fieldId;
+    uint256 fieldId;
     ProtoBufRuntime.WireType wireType;
-    uint bytesRead;
-    uint offset = p;
-    uint pointer = p;
-    while(pointer < offset+sz) {
+    uint256 bytesRead;
+    uint256 offset = p;
+    uint256 pointer = p;
+    while (pointer < offset + sz) {
       (fieldId, wireType, bytesRead) = ProtoBufRuntime._decode_key(pointer, bs);
       pointer += bytesRead;
-      if(fieldId == 1) {
+      if (fieldId == 1) {
         pointer += _read_instrumentTerminationTimestamp(pointer, bs, r, counters);
       }
-      else if(fieldId == 2) {
+      else if (fieldId == 2) {
         pointer += _read_instrumentOverrideTimestamp(pointer, bs, r, counters);
       }
-      else if(fieldId == 3) {
+      else if (fieldId == 3) {
         pointer += _read_brokerAddress(pointer, bs, r, counters);
       }
-      else if(fieldId == 4) {
+      else if (fieldId == 4) {
         pointer += _read_supportMakerWhitelist(pointer, bs, r, counters);
       }
-      else if(fieldId == 5) {
+      else if (fieldId == 5) {
         pointer += _read_supportTakerWhitelist(pointer, bs, r, counters);
       }
       
       else {
         if (wireType == ProtoBufRuntime.WireType.Fixed64) {
-          uint size;
+          uint256 size;
           (, size) = ProtoBufRuntime._decode_fixed64(pointer, bs);
           pointer += size;
         }
         if (wireType == ProtoBufRuntime.WireType.Fixed32) {
-          uint size;
+          uint256 size;
           (, size) = ProtoBufRuntime._decode_fixed32(pointer, bs);
           pointer += size;
         }
         if (wireType == ProtoBufRuntime.WireType.Varint) {
-          uint size;
+          uint256 size;
           (, size) = ProtoBufRuntime._decode_varint(pointer, bs);
           pointer += size;
         }
         if (wireType == ProtoBufRuntime.WireType.LengthDelim) {
-          uint size;
+          uint256 size;
           (, size) = ProtoBufRuntime._decode_lendelim(pointer, bs);
           pointer += size;
         }
@@ -109,16 +112,21 @@ library InstrumentParameters {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_instrumentTerminationTimestamp(uint p, bytes memory bs, Data memory r, uint[6] memory counters) internal pure returns (uint) {
+  function _read_instrumentTerminationTimestamp(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[6] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (uint256 x, uint sz) = ProtoBufRuntime._decode_sol_uint256(p, bs);
-    if(isNil(r)) {
+    (uint256 x, uint256 sz) = ProtoBufRuntime._decode_sol_uint256(p, bs);
+    if (isNil(r)) {
       counters[1] += 1;
     } else {
       r.instrumentTerminationTimestamp = x;
-      if(counters[1] > 0) counters[1] -= 1;
+      if (counters[1] > 0) counters[1] -= 1;
     }
     return sz;
   }
@@ -131,16 +139,21 @@ library InstrumentParameters {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_instrumentOverrideTimestamp(uint p, bytes memory bs, Data memory r, uint[6] memory counters) internal pure returns (uint) {
+  function _read_instrumentOverrideTimestamp(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[6] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (uint256 x, uint sz) = ProtoBufRuntime._decode_sol_uint256(p, bs);
-    if(isNil(r)) {
+    (uint256 x, uint256 sz) = ProtoBufRuntime._decode_sol_uint256(p, bs);
+    if (isNil(r)) {
       counters[2] += 1;
     } else {
       r.instrumentOverrideTimestamp = x;
-      if(counters[2] > 0) counters[2] -= 1;
+      if (counters[2] > 0) counters[2] -= 1;
     }
     return sz;
   }
@@ -153,16 +166,21 @@ library InstrumentParameters {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_brokerAddress(uint p, bytes memory bs, Data memory r, uint[6] memory counters) internal pure returns (uint) {
+  function _read_brokerAddress(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[6] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (address x, uint sz) = ProtoBufRuntime._decode_sol_address(p, bs);
-    if(isNil(r)) {
+    (address x, uint256 sz) = ProtoBufRuntime._decode_sol_address(p, bs);
+    if (isNil(r)) {
       counters[3] += 1;
     } else {
       r.brokerAddress = x;
-      if(counters[3] > 0) counters[3] -= 1;
+      if (counters[3] > 0) counters[3] -= 1;
     }
     return sz;
   }
@@ -175,16 +193,21 @@ library InstrumentParameters {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_supportMakerWhitelist(uint p, bytes memory bs, Data memory r, uint[6] memory counters) internal pure returns (uint) {
+  function _read_supportMakerWhitelist(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[6] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (bool x, uint sz) = ProtoBufRuntime._decode_bool(p, bs);
-    if(isNil(r)) {
+    (bool x, uint256 sz) = ProtoBufRuntime._decode_bool(p, bs);
+    if (isNil(r)) {
       counters[4] += 1;
     } else {
       r.supportMakerWhitelist = x;
-      if(counters[4] > 0) counters[4] -= 1;
+      if (counters[4] > 0) counters[4] -= 1;
     }
     return sz;
   }
@@ -197,16 +220,21 @@ library InstrumentParameters {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_supportTakerWhitelist(uint p, bytes memory bs, Data memory r, uint[6] memory counters) internal pure returns (uint) {
+  function _read_supportTakerWhitelist(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[6] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (bool x, uint sz) = ProtoBufRuntime._decode_bool(p, bs);
-    if(isNil(r)) {
+    (bool x, uint256 sz) = ProtoBufRuntime._decode_bool(p, bs);
+    if (isNil(r)) {
       counters[5] += 1;
     } else {
       r.supportTakerWhitelist = x;
-      if(counters[5] > 0) counters[5] -= 1;
+      if (counters[5] > 0) counters[5] -= 1;
     }
     return sz;
   }
@@ -221,7 +249,7 @@ library InstrumentParameters {
    */
   function encode(Data memory r) internal pure returns (bytes memory) {
     bytes memory bs = new bytes(_estimate(r));
-    uint sz = _encode(r, 32, bs);
+    uint256 sz = _encode(r, 32, bs);
     assembly {
       mstore(bs, sz)
     }
@@ -236,20 +264,48 @@ library InstrumentParameters {
    * @param bs The bytes array to be decoded
    * @return The number of bytes encoded
    */
-  function _encode(Data memory r, uint p, bytes memory bs)
-      internal pure returns (uint) {
-    uint offset = p;
-    uint pointer = p;
+  function _encode(Data memory r, uint256 p, bytes memory bs)
+    internal 
+    pure 
+    returns (uint) 
+  {
+    uint256 offset = p;
+    uint256 pointer = p;
     
-    pointer += ProtoBufRuntime._encode_key(1, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      1, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_uint256(r.instrumentTerminationTimestamp, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(2, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      2, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_uint256(r.instrumentOverrideTimestamp, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(3, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      3, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_address(r.brokerAddress, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(4, ProtoBufRuntime.WireType.Varint, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      4, 
+      ProtoBufRuntime.WireType.Varint, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_bool(r.supportMakerWhitelist, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(5, ProtoBufRuntime.WireType.Varint, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      5, 
+      ProtoBufRuntime.WireType.Varint, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_bool(r.supportTakerWhitelist, pointer, bs);
     return pointer - offset;
   }
@@ -262,18 +318,21 @@ library InstrumentParameters {
    * @param bs The bytes array to be decoded
    * @return The number of bytes encoded
    */
-  function _encode_nested(Data memory r, uint p, bytes memory bs)
-      internal pure returns (uint) {
+  function _encode_nested(Data memory r, uint256 p, bytes memory bs)
+    internal 
+    pure 
+    returns (uint) 
+  {
     /**
      * First encoded `r` into a temporary array, and encode the actual size used.
      * Then copy the temporary array into `bs`.
      */
-    uint offset = p;
-    uint pointer = p;
+    uint256 offset = p;
+    uint256 pointer = p;
     bytes memory tmp = new bytes(_estimate(r));
-    uint tmpAddr = ProtoBufRuntime.getMemoryAddress(tmp);
-    uint bsAddr = ProtoBufRuntime.getMemoryAddress(bs);
-    uint size = _encode(r, 32, tmp);
+    uint256 tmpAddr = ProtoBufRuntime.getMemoryAddress(tmp);
+    uint256 bsAddr = ProtoBufRuntime.getMemoryAddress(bs);
+    uint256 size = _encode(r, 32, tmp);
     pointer += ProtoBufRuntime._encode_varint(size, pointer, bs);
     ProtoBufRuntime.copyBytes(tmpAddr + 32, bsAddr + pointer, size);
     pointer += size;
@@ -286,8 +345,10 @@ library InstrumentParameters {
    * @dev The estimator for a struct
    * @return The number of bytes encoded in estimation
    */
-  function _estimate(Data memory /* r */) internal pure returns (uint) {
-    uint e;
+  function _estimate(
+    Data memory /* r */
+  ) internal pure returns (uint) {
+    uint256 e;
     e += 1 + 35;
     e += 1 + 35;
     e += 1 + 23;
